Monday fixes 3 (#55)
* fix: container_only visibility

* add: debounce to most submits

* add: debounce to all relevant functions

* add: try-except to EventAdmin

* add: debounce to Modal.vue

* fix: schoolGroup was not required in Event Model

* fix: email sent on user info change

* fix: email send bug

* add: search & list filters

* remove: excel upload on mobile devices

* add: django server gitpod visibility

* add: empty slug console.error msg

* fix: numeric school grades

* add: placeholder msg & fix: website validation rule in SchoolDetails

* add: domain choices instead of free text & adjust fe

* add: migrations

+++ client/src/api/vendorProgram.js
@@ -14,6 +14,8 @@ const vendorProgram = {
     return axios.get(`${GET_VENDOR_PROGRAM_LIST_API_URL}${slug}/`)
   },
   getProgramMediaList(programSlug) {
+    if (!programSlug)
+      throw "getProgramMediaList: received empty slug"
     return axios.get(GET_PROGRAM_MEDIA_LIST_API_URL, {
       params: {
         activity__slug: programSlug,
@@ -21,6 +23,8 @@ const vendorProgram = {
     })
   },
   deleteProgramMedia(mediaSlug) {
+    if (!mediaSlug)
+      throw "deleteProgramMedia: received empty slug"
     return axios.delete(`${DELETE_PROGRAM_MEDIA_API_URL}${mediaSlug}/`)
   },
   createProgramMedia(data) {
@@ -37,12 +41,11 @@ const vendorProgram = {
     return axios.post(CREATE_VENDOR_PROGRAM_API_URL, data)
   },
   updateProgram(programSlug, data) {
-    return axios.patch(
-      `${UPDATE_VENDOR_PROGRAM_API_URL}${programSlug}/`,
-      data
-    )
+    if (!programSlug) throw "updateProgram: received empty slug"
+    return axios.patch(`${UPDATE_VENDOR_PROGRAM_API_URL}${programSlug}/`, data)
   },
   deleteProgram(programSlug) {
+    if (!programSlug) throw "deleteProgram: received empty slug"
     return axios.delete(`${DELETE_VENDOR_PROGRAM_API_URL}${programSlug}/`)
   },
 }
